Move snapshot execution/scheduling out of root of snapshots module and into a submodule
